Fix the typo and move the chartCanvas out as one component.
